sources: enable 7z, bzr, hg, svn, zip for non-linux
Zip uses the python zipfile implementation.

The remainder will fail if the equivalent tools are not in the
user's path on the host. That error should be more obvious
than the "KeyError" uses are currently getting when using
remote-build.  Future work should update these source handlers
to use file_utils.get_host_tool_path() to perform this check
prior to executing the command.

LP: #1898877

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>